Revert the player to having no receiver
I understand now that receiver exists just in case you prefer a declarative interface which updates the component whenever the container is re-rendered.  This in turn happens when the container writes to its state.

I'm no longer convinced we need this behaviour after all, and so this branch looks at a player with no receiver (which thus must be updated imperatively).